Makefile updated and parser.c issue solved
Updated makefile to fix the new compiler clang. Now the flag -lreadline has been removed
from the general flag rule, to be used just when the complete program is compiled.
Another issue fixed when the AST has an identifier in the root. The problem was that
I was not considering that case.