Global variable
I created a global variable that manage signal. It is declarated in bash.h and define in main.c. I created a signal directory that contain a single file  with the only file that
I created. In addiction, I fixed an error that save in the history of command when the command sentence was empty, when bash do not save in this case. Icreate a if case to
add_history funtion to fixed.