Another change to assignments
The assignments had problems when they were in the beginning of the
input line. In that case, the parser didn't work fine and it didn't
enter to the right function.

The problem was solved changing the ft_assignment function. In the
condition for assignments, I changed the !list condition fon !*list.