Norminette changes and issue with env variables
A issue with dollar sustitution has been solved. The problem happened
because the get_variable function didn't finish the strings with null.
This was doing the minishell to have undefined behavior.

Also a lot of norminette errors have been solved among the functions
from expanser. Various new files has been created for this.