Input with spaces solved again
The problem that caused a segmentation fault when the input were filled with spaces
returned today. I made a correction in the skip_spaces function, in order to prevent
this segmentation faults to occur again. Now the tokenizer should send always a EOF token
when the input is composed for spaces.