Misterious open solved, l"s" solved
The misterious open has been finally solved. The problem was that the
child proccess was impersonating the parent process, and the parent was
just waiting the child to finish.

Also there was another problem related with the command l"s", when the
different nodes were concanated, the old nodes still remained.