Pipes solved, Quotes solved, CtrlC segfault solved
The problem with the pipes has been solved, a string was freed before
been used in command_args_builder.

The segfault with unclosed quotes have been solved as well. The program
was entering in the ft_undercore() function when the tokens were NULL.

The function ft_underscore_arg() was causing another problems when
minishell->underscore didn't exist, such as the segmentation fault in
the first line when pressing Ctrl + C with some prompt.